requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==41.0.7
cycler==0.12.1
fairlearn==0.13.0
flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
fonttools==4.60.2
google-ai-generativelanguage==0.6.4
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.5.4
googleapis-common-protos==1.73.0
grpcio==1.59.3
httplib2==0.31.2
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
markupsafe==3.0.3
matplotlib==3.9.4
mysql-connector-python==8.3.0
narwhals==2.20.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==26.1
pandas==2.3.3
pillow==11.3.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
reportlab==4.0.8
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
werkzeug==3.1.6
zipp==3.23.0
